Silence a few warnings from `commitData`
When creating a new widget from delegates in `createEditor`, we set
their values, then return the new widget. If we connect to `commitData`
too soon, it will be called when the widget changes values, but it has
not returned yet from `createEditor`, so the view will complain
repeatedly about it. Mitigate it blocking signals and connecting later.